requirements.txt
python-dotenv
elevenlabs
twilio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
Flask
pytz
